requirements.txt
altgraph==0.10.2
asn1crypto==0.24.0
bdist-mpkg==0.5.0
bonjour-py==0.3
CacheControl==0.12.6
cachetools==3.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
enum34==1.1.6
firebase-admin==3.2.1
Flask==1.1.1
future==0.17.1
futures==3.3.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
macholib==1.5.1
MarkupSafe==1.1.1
matplotlib==1.3.1
modulegraph==0.10.4
msgpack==1.0.0
numpy==1.8.0rc1
protobuf==3.11.3
py2app==0.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.23.0
rsa==4.0
scipy==0.13.0b1
selenium==3.13.0
six==1.12.0
uritemplate==3.0.1
urllib3==1.25.8
vboxapi==1.0
virtualenv==16.0.0
Werkzeug==0.15.5
xattr==0.6.4
